containerize servers
