+++ .gitignore
@@ -1,10 +1,6 @@
 *.o
 *.pb.cc
 *.pb.h
-supplier_server
-vendor_server
-finder_server
-client
 cmake/
 cmake_externalproject
 cocoapods
